Split BT ports files to break cyclic dependencies
