+++ src/as2fm/scxml_converter/scxml_entries/__init__.py
@@ -3,13 +3,12 @@
 from .scxml_base import ScxmlBase  # noqa: F401
 from .utils import CallbackType  # noqa: F401
 from .bt_utils import RESERVED_BT_PORT_NAMES  # noqa: F401
-from .scxml_bt_ports import (  # noqa: F401
+from .scxml_bt_port_declaration import (  # noqa: F401
     BtInputPortDeclaration,
     BtPortDeclarations,
     BtOutputPortDeclaration,
-    BtGetValueInputPort,
-    BtSetValueOutputPort,
 )  # noqa: F401
+from .scxml_bt_in_port import BtGetValueInputPort  # noqa: F401
 from .scxml_param import ScxmlParam  # noqa: F401
 from .scxml_ros_field import RosField  # noqa: F401
 from .scxml_data import ScxmlData  # noqa: F401
@@ -21,6 +20,7 @@ from .scxml_executable_entries import (  # noqa: F401
     ScxmlExecutionBody,
     EventsToAutomata,
 )  # noqa: F401
+from .scxml_bt_out_port import BtSetValueOutputPort  # noqa: F401
 from .scxml_executable_entries import (  # noqa: F401
     execution_body_from_xml,
     as_plain_execution_body,
