+++ src/as2fm/scxml_converter/scxml_entries/scxml_bt_port_declaration.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 
 """
-SCXML entries related to Behavior Trees' Ports.
+SCXML entries related to Behavior Trees' Ports declaration.
 """
 
 from typing import Union
@@ -109,74 +109,6 @@ class BtOutputPortDeclaration(ScxmlBase):
         xml_bt_in_port = ET.Element(
             BtOutputPortDeclaration.get_tag_name(), {"key": self._key, "type": self._type}
         )
-        return xml_bt_in_port
-
-
-class BtGetValueInputPort(ScxmlBase):
-    """
-    Get the value of an input port in a bt plugin.
-    """
-
-    @staticmethod
-    def get_tag_name() -> str:
-        return "bt_get_input"
-
-    @staticmethod
-    def from_xml_tree(xml_tree: ET.Element) -> "BtGetValueInputPort":
-        assert_xml_tag_ok(BtGetValueInputPort, xml_tree)
-        key_str = get_xml_argument(BtGetValueInputPort, xml_tree, "key")
-        return BtGetValueInputPort(key_str)
-
-    def __init__(self, key_str: str):
-        self._key = key_str
-
-    def check_validity(self) -> bool:
-        return is_non_empty_string(BtGetValueInputPort, "key", self._key)
-
-    def get_key_name(self) -> str:
-        return self._key
-
-    def as_plain_scxml(self, _) -> ScxmlBase:
-        # This is discarded in the to_plain_scxml_and_declarations method from ScxmlRoot
-        raise RuntimeError("Error: SCXML BT Port value getter cannot be converted to plain SCXML.")
-
-    def as_xml(self) -> ET.Element:
-        assert self.check_validity(), "Error: SCXML BT Input Port: invalid parameters."
-        xml_bt_in_port = ET.Element(BtGetValueInputPort.get_tag_name(), {"key": self._key})
-        return xml_bt_in_port
-
-
-class BtSetValueOutputPort(ScxmlBase):
-    """
-    Get the value of an input port in a bt plugin.
-    """
-
-    @staticmethod
-    def get_tag_name() -> str:
-        return "bt_set_output"
-
-    @staticmethod
-    def from_xml_tree(xml_tree: ET.Element) -> "BtSetValueOutputPort":
-        assert_xml_tag_ok(BtSetValueOutputPort, xml_tree)
-        key_str = get_xml_argument(BtSetValueOutputPort, xml_tree, "key")
-        return BtSetValueOutputPort(key_str)
-
-    def __init__(self, key_str: str):
-        self._key = key_str
-
-    def check_validity(self) -> bool:
-        return is_non_empty_string(BtSetValueOutputPort, "key", self._key)
-
-    def get_key_name(self) -> str:
-        return self._key
-
-    def as_plain_scxml(self, _) -> ScxmlBase:
-        # This is discarded in the to_plain_scxml_and_declarations method from ScxmlRoot
-        raise RuntimeError("Error: SCXML BT Port value setter cannot be converted to plain SCXML.")
-
-    def as_xml(self) -> ET.Element:
-        assert self.check_validity(), "Error: SCXML BT Input Port: invalid parameters."
-        xml_bt_in_port = ET.Element(BtSetValueOutputPort.get_tag_name(), {"key": self._key})
         return xml_bt_in_port
